Updated tables for move generation

+++ src/square.py
@@ -13,9 +13,8 @@ class Square(object):
 
     @classmethod
     def from_position(cls, r, f):
-        return cls((r.value<<3) | f.value) # 8*rank + file
+        return cls((r.value << np.uint8(3)) | f.value) # 8*rank + file
 
     def to_bitboard(self):
         return np.uint64(1) << self.index
-
